split tracking to its submodule
